:lipstick: Separated CSS into a dedicated directory
