+++ frontend/src/css/InfoPanel.css
